+++ main/color.h
@@ -8,7 +8,7 @@
 #define color_orange  0xffAA00
 
 uint32_t color_pack(uint8_t red, uint8_t green, uint8_t blue);
-uint8_t color_getRed(uint8_t);
-uint8_t color_getGreen(uint8_t);
-uint8_t color_getBlue(uint8_t);
+uint8_t color_getRed(uint32_t);
+uint8_t color_getGreen(uint32_t);
+uint8_t color_getBlue(uint32_t);
 uint8_t color_getGray(uint32_t);
